# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Stellar DLMM Liquidity Protocol API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Token {
  symbol: string;
  name: string;
  address: string;
  decimals: number;
  price: number;
  priceChange24h?: number;
  logoUrl: string;
}

/**
 * dlmm = our on-chain DLMM contract; amm = native Stellar DEX liquidity pool (aggregated from Horizon)
 */
export type PoolCategory = typeof PoolCategory[keyof typeof PoolCategory];


export const PoolCategory = {
  dlmm: 'dlmm',
  amm: 'amm',
} as const;

export interface Pool {
  id: string;
  /** dlmm = our on-chain DLMM contract; amm = native Stellar DEX liquidity pool (aggregated from Horizon) */
  category: PoolCategory;
  tokenX: Token;
  tokenY: Token;
  tvl: number;
  volume24h: number;
  fees24h: number;
  apr: number;
  binStep: number;
  activeBinId: number;
  currentPrice?: number;
  fee?: number;
  reserveX?: number;
  reserveY?: number;
  /** Total LP shares/pool shares outstanding (AMM pools from Horizon) */
  totalShares?: number;
  /** Link to view an AMM pool on stellar.expert */
  externalUrl?: string;
  /** False when 24h volume/APR is not indexed for this pool (shown as — in UI) */
  volumeAvailable?: boolean;
}

export type PoolDetailCategory = typeof PoolDetailCategory[keyof typeof PoolDetailCategory];


export const PoolDetailCategory = {
  dlmm: 'dlmm',
  amm: 'amm',
} as const;

export interface PoolDetail {
  id: string;
  category: PoolDetailCategory;
  tokenX: Token;
  tokenY: Token;
  tvl: number;
  volume24h: number;
  fees24h: number;
  apr: number;
  binStep: number;
  activeBinId: number;
  currentPrice: number;
  fee: number;
  reserveX: number;
  reserveY: number;
  totalBins: number;
  contractAddress?: string;
  totalShares?: number;
  externalUrl?: string;
  volumeAvailable?: boolean;
}

export interface Bin {
  binId: number;
  price: number;
  liquidityX: number;
  liquidityY: number;
  isActive: boolean;
  totalLiquidity?: number;
}

export interface PoolStatsPoint {
  timestamp: string;
  tvl: number;
  volume: number;
  fees: number;
}

export interface PoolStats {
  poolId: string;
  period: string;
  data: PoolStatsPoint[];
}

export interface ProtocolSummary {
  totalTvl: number;
  totalVolume24h: number;
  totalFees24h: number;
  totalPools: number;
  totalTransactions24h: number;
  tvlChange24h?: number;
  volumeChange24h?: number;
}

export interface SwapQuoteInput {
  tokenInAddress: string;
  tokenOutAddress: string;
  amountIn: number;
  slippageTolerance?: number;
}

export interface SwapQuote {
  tokenIn: Token;
  tokenOut: Token;
  amountIn: number;
  amountOut: number;
  priceImpact: number;
  minimumReceived: number;
  fee: number;
  route: string[];
  executionPrice?: number;
  binsTraversed?: number;
}

export interface SwapRouteInput {
  tokenInAddress: string;
  tokenOutAddress: string;
  amountIn: number;
}

export interface RouteHop {
  poolId: string;
  tokenIn: Token;
  tokenOut: Token;
  amountIn: number;
  amountOut: number;
}

export interface SwapRoute {
  hops: RouteHop[];
  totalAmountOut: number;
  totalPriceImpact: number;
  totalFee: number;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  swap: 'swap',
  add_liquidity: 'add_liquidity',
  remove_liquidity: 'remove_liquidity',
} as const;

export interface Transaction {
  id: string;
  type: TransactionType;
  poolId: string;
  address: string;
  timestamp: string;
  txHash: string;
  tokenXAmount?: number;
  tokenYAmount?: number;
  valueUsd?: number;
  binsAffected?: number[];
}

export type PositionStrategy = typeof PositionStrategy[keyof typeof PositionStrategy];


export const PositionStrategy = {
  spot: 'spot',
  curve: 'curve',
  bid_ask: 'bid_ask',
} as const;

export interface Position {
  id: string;
  poolId: string;
  pool?: Pool;
  address: string;
  binRangeLow: number;
  binRangeHigh: number;
  /** The single bin this position occupies (DLMM positions are per-bin) */
  binId?: number;
  /** LP shares held in the bin */
  shares?: number;
  liquidityX: number;
  liquidityY: number;
  valueUsd: number;
  unrealizedFees: number;
  strategy?: PositionStrategy;
}

export type ListPoolsParams = {
sortBy?: ListPoolsSortBy;
search?: string;
};

export type ListPoolsSortBy = typeof ListPoolsSortBy[keyof typeof ListPoolsSortBy];


export const ListPoolsSortBy = {
  tvl: 'tvl',
  volume24h: 'volume24h',
  apr: 'apr',
  fees24h: 'fees24h',
} as const;

export type ListTransactionsParams = {
poolId?: string;
address?: string;
type?: ListTransactionsType;
limit?: number;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType];


export const ListTransactionsType = {
  swap: 'swap',
  add_liquidity: 'add_liquidity',
  remove_liquidity: 'remove_liquidity',
} as const;