# lib/api-zod/src/generated/types/poolDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Stellar DLMM Liquidity Protocol API
 * OpenAPI spec version: 0.1.0
 */
import type { PoolDetailCategory } from './poolDetailCategory';
import type { Token } from './token';

export interface PoolDetail {
  id: string;
  category: PoolDetailCategory;
  tokenX: Token;
  tokenY: Token;
  tvl: number;
  volume24h: number;
  fees24h: number;
  apr: number;
  binStep: number;
  activeBinId: number;
  currentPrice: number;
  fee: number;
  reserveX: number;
  reserveY: number;
  totalBins: number;
  contractAddress?: string;
  totalShares?: number;
  externalUrl?: string;
  volumeAvailable?: boolean;
}